🚚 Move IMAP#responses tests to their own file
`test_imap.rb` is huge!

`wait_for_response_count` was moved into TestCase (in test/lib/helper).
Perhaps this should be moved into a shared module or an IMAPTestCase
class (subclass of TestCase) instead?
